+++ examples/resources/morpheus_guidance_setting/import.sh
@@ -1,0 +1,1 @@
+terraform import morpheus_guidance_setting.tf_example_guidance_config 1