+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 NOTES:
 
-* The `morpheus-sdk` dependcy has been upgraded to version 0.2.10.
+* The `morpheus-sdk` dependcy has been upgraded to version 0.3.0.
+* The `terraform-plugin-docs` dependcy has been upgraded to version 0.14.1.
+* The `terraform-plugin-sdk` dependcy has been upgraded to version 2.25.0.
 * Add label support for additional Morpheus resources.
 * Update inputs to support additional configuration parameters (i.e. - editable, verify pattern, etc).
 
@@ -10,6 +12,7 @@ FEATURES:
 
 * **New Data Source:** `morpheus_vro_workflow`
 * **New Resource:** `morpheus_active_directory_identity_source`
+* **New Resource:** `morpheus_guidance_setting`
 * **New Resource:** `morpheus_vro_integration`
 * **New Resource:** `morpheus_vro_task`